contract: move struct PushCommand to phala-types
