+++ common/types/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2018"
 
 [dependencies]
 serde = { version = "1.0.101", default-features = false, optional = true }
-codec = { package = "parity-scale-codec", version = "2.1", default-features = false }
+codec = { package = "parity-scale-codec", version = "2.1", default-features = false, features = ["full"] }
 sp-core = { version = "3.0.0", default-features = false, path = "../../substrate/primitives/core" }
 sp-runtime = { default-features = false, path = "../../substrate/primitives/runtime", optional = true }
